requirements.txt
pip==25.0.1

# Flask and related packages
Flask==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-Limiter==2.7.0
Flask-CORS==3.0.10
Jinja2==3.1.2  # Added Jinja2 for template rendering

# WebSockets & Real-time communication (✅ Compatible versions)
Flask-SocketIO==5.3.6  # WebSockets for Flask
eventlet==0.33.3  # Async networking library (Required for Flask-SocketIO)
python-engineio==4.11.2
python-socketio==5.12.1


# Authentication
PyJWT==2.6.0
bcrypt

# HTTP requests
requests==2.31.0

# For managing environment variables
python-dotenv==1.0.0

# Database driver for PostgreSQL
psycopg2-binary==2.9.6  # ✅ Required for PostgreSQL
cryptography==42.0.5  # Use the latest stable version
redis==5.0.1  # Use the latest stable version

# For linting and formatting (optional)
pylint==2.17.5
flake8==6.0.0
black==23.1.0

# YAML support
PyYAML==6.0.1
